backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.5
motor>=3.3.0
dnspython>=2.4.0
python-dotenv>=1.0.0
